refactor: restructure page and template components for improved organization
